Remove inProgressReaction, add slackTimeout config
The slackTimeout seems like a good thing to have. Both it and successReaction
are passed into the SlackClient.

Removing inProgressReaction because, after thinking on it more, I realized
that an in-memory object whose keys are message IDs (composed of the
`channel:` and `timestamp:` members from the event message) should suffice.